requirements.txt
streamlit
pandas
numpy
altair
plotly
jira
mysql-connector-python
openpyxl